Split reanimatedWrapper
