+++ src/handlers/gestures/reanimatedWrapper.web.ts
@@ -1,0 +1,1 @@
+export * from './reanimatedWrapper.native';
